requirements.txt
pyjwt
python-dotenv
flask
flasgger
flask_sqlalchemy
flask-cors
flask-jwt-extended
flask-swagger-ui
pymysql